requirements.txt
python-dotenv
streamlit
playwright
feedparser
requests
beautifulsoup4
